Functions tests update
